@@ -174,7 +174,7 @@ class _DateRangePickerState extends State<DateRangePicker> {
                   shape: BoxShape.circle,
                 ),
                 todayDecoration: BoxDecoration(
-                  color: AppTheme.primaryColor.withOpacity(0.2),
+                  color: AppTheme.primaryColor.withValues(alpha:0.2),
                   shape: BoxShape.circle,
                 ),
                 outsideDaysVisible: false,
