@@ -1,48 +1,63 @@
 class ApiConfig {
   static const String devBaseUrl = 'http://127.0.0.1:8000/api';
-  static const String prodBaseUrl = 'https://your-production-url.com/api';
+  static const String prodBaseUrl = 'http://63.177.98.17/api';
 
-  static const bool isProduction = false;
+  static const bool isProduction = true;
 
   static String get baseUrl => isProduction ? prodBaseUrl : devBaseUrl;
 
-  static String getImageUrl(String imageUrl) => '$baseUrl/images/$imageUrl';
-
   //Authorization
-  static Uri getRegisterUrl() => Uri.parse('$baseUrl/register');
-  static Uri getLoginUrl() => Uri.parse('$baseUrl/login');
-  static Uri getLogoutUrl() => Uri.parse('$baseUrl/logout');
-  static Uri getRefreshTokenUrl() => Uri.parse('$baseUrl/token/refresh');
+  static Uri register() => Uri.parse('$baseUrl/register');
+  static Uri login() => Uri.parse('$baseUrl/login');
+  static Uri logout() => Uri.parse('$baseUrl/logout');
+  static Uri refreshToken() => Uri.parse('$baseUrl/token/refresh');
 
   //User
-  static Uri getUserUrl() => Uri.parse('$baseUrl/user');
-  static Uri getUserByIdAdminUrl(String id) =>
-      Uri.parse('$baseUrl/admin/user/$id');
-  static Uri getUsersAdminUrl() => Uri.parse('$baseUrl/admin/users');
-  static Uri getUserChangePasswordUrl() =>
+  static Uri userDetails() => Uri.parse('$baseUrl/user/details');
+  static Uri changePassword() =>
       Uri.parse('$baseUrl/user/change_password');
-  static Uri getDeleteUserByIdAdminUrl(String id) =>
-      Uri.parse('${getUserByIdAdminUrl(id)}/delete');
-  static Uri getUpdateUserRolesByIdAdminUrl(String id) =>
-      Uri.parse('${getUserByIdAdminUrl(id)}/update_roles');
+  static Uri userById(String id) =>
+      Uri.parse('$baseUrl/admin/user/$id');
+  static Uri users() => Uri.parse('$baseUrl/admin/users');
+  static Uri userDetailsById(String id) =>
+      Uri.parse('${userById(id)}/details');
+  static Uri deleteUserById(String id) =>
+      Uri.parse('${userById(id)}/delete');
+  static Uri updateRolesById(String id) =>
+      Uri.parse('${userById(id)}/update_roles');
 
-  //Meals
-  static Uri getMealPlansUrl() => Uri.parse('$baseUrl/meal_plans');
-  static Uri deleteMealPlanUrl(String id) => Uri.parse('$baseUrl/admin/meal_plans/$id/delete');
-  static Uri updateMealPlanUrl(String id) => Uri.parse('$baseUrl/admin/meal_plans/$id/update');
-  static Uri createMealPlanUrl() => Uri.parse('$baseUrl/admin/meal_plans/new');
-  static Uri getMealCategoriesUrl() => Uri.parse('$baseUrl/categories');
-  static Uri getMealsUrl(String? category) => category == null
+  //Meals Plans
+  static Uri mealPlans() => Uri.parse('$baseUrl/meal_plan');
+  static Uri newMealPlan() => Uri.parse('$baseUrl/admin/meal_plan');
+  static Uri updateMealPlan(String id) =>
+      Uri.parse('$baseUrl/admin/meal_plan/$id');
+  static Uri deleteMealPlan(String id) =>
+      Uri.parse('$baseUrl/admin/meal_plan/$id/delete');
+
+  //Meal
+  static Uri meals(String? category) => category == null
       ? Uri.parse('$baseUrl/meals')
       : Uri.parse('$baseUrl/meals/$category');
+  static Uri newMeal() => Uri.parse('$baseUrl/admin/meal');
+  static Uri updateMeal(String id) => Uri.parse('$baseUrl/admin/meal/$id');
+  static Uri deleteMeal(String id) =>
+      Uri.parse('$baseUrl/admin/meal/$id/delete');
+
+  //Categories
+  static Uri categories() => Uri.parse('$baseUrl/categories');
+  static Uri newCategory() => Uri.parse('$baseUrl/admin/category');
+  static Uri updateCategory(String id) =>
+      Uri.parse('$baseUrl/admin/category/$id');
+  static Uri deleteCategory(String id) =>
+      Uri.parse('$baseUrl/admin/category/$id/delete');
 
   //Orders
-  static Uri getUserOrdersUrl() => Uri.parse('$baseUrl/user/orders');
-  static Uri getNewOrderUrl() => Uri.parse('$baseUrl/order/new');
-  static Uri getUserOrdersByIdAdminUrl(String id) =>
-      Uri.parse('${getUserByIdAdminUrl(id)}/orders');
-  static Uri getDeleteOrderByIdAdminUrl(String id) =>
+  static Uri orders() => Uri.parse('$baseUrl/user/orders');
+  static Uri newOrder() => Uri.parse('$baseUrl/order');
+  static Uri ordersByUserId(String id) =>
+      Uri.parse('${userById(id)}/orders');
+  static Uri deleteOrder(String id) =>
       Uri.parse('$baseUrl/admin/order/$id/delete');
-  static Uri getUpdateOrderStatusUrl(String id) =>
+  static Uri updateOrderStatus(String id) =>
       Uri.parse('$baseUrl/admin/order/$id/update_status');
 }
