@@ -33,7 +33,7 @@ class AppTheme {
       cardTheme: CardTheme(
         color: cardColor,
         elevation: 2,
-        margin: const EdgeInsets.symmetric(vertical: 8),
+        margin: const EdgeInsets.symmetric(vertical: AppTheme.defaultPadding/2),
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(cardRadius),
         ),
